Reuse persistence upload helper on overlay page
